[Asterisk-Dev] better pattern matcher
well, for the beginning I added support to continue to match after a "."
to be able to add extensions after the dialed number and not only before. :)

e.g "_0N.8500" matches my voicemail extension for all dialed numbers.

I'll commit this short patch on Tuesday, because I have to do more
testing and the weekend is free.

But maybe I add quantifiers also and support
   "." for any char,
   "+" for "one or more"
   "*" for zero or more chars.
This will break existing extensions.

BTW: character classes (alphanumerics) are already supported by [].
[a-z0-9] for example.

I'm not good in interpreting licenses. Can somebody look at the PCRE
license. I always use this in my projects and it should be able to use
this commercially.

http://lists.digium.com/pipermail/asterisk-dev/2003-June/000847.html